+++ aura/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - The `--skipdepcheck` flag to skip all dependency solving. Combine this with
   `--hotedit` to avoid discrepancies in custom dependency listings.
+- `--skippgpcheck` is now available to pass through to `makepkg`.
 
 #### Changed